🎨 Set up Elixir formatter
Add a `test.all` alias that runs the unit tests and checks that all files are formatted.  Make this the default command for the `server_test` container.

Files can be formatted with `docker-compose exec server mix format` as needed.

In-editor formatting seems to work OK, but that might be because I have a version of Elixir installed on my workstation.

There doesn't seem to be a way to get the Elixir language server to run from the Docker container, so a new strategy will be required (future PR).