🏷️ Add type-checking for server code
Add dialyxir for doing type-checking on the server code.

For now, run it in the `server_test` container, which is in the `test` mix environment.

Fix an incorrect typespec, add `ex_unit` as an explicit dependency (see https://elixirforum.com/t/help-with-dialyzer-output/15202/5), and add an ignore for a warning that only shows up because I don't have any routes yet (see https://github.com/phoenixframework/phoenix/issues/2750).

I'm not sure the default paths for the PLT files will cache well in either Docker containers or CI, but will explore further later.  For now, it seems to work OK.