🐳 Restore server_test container
On second thought, it seems better to have a pre-configured container for running the tests and related tasks.

While we still need a local install for editor integration, we can do everything else in containers.

The restored `server_test` container is configured differently than before.  It uses the exact same `Dockerfile` as the `db_setup`/`server` containers, but runs with `MIX_ENV=test`.

In order to cache dialyzer's PLT files, I've moved them to `server/_plts` and bind-mounted them into the container.

Also:
- Removed the now-obsolete `.db` directory
- Modified bind-mounts to be read-only where possible.