🔧 Add pre-commit hooks to check lint/format
Use husky and lint-staged to define pre-commit hooks.

The hooks check formatting and linting, auto-fixing where possible.
Unfortunately, `mix credo` doesn't take a list of files on the command
line, so for now I'm not including it in the hook.

lint-staged recently added support for handling partially-staged files,
so I'm testing that out with auto-fixing and `git add`ing.  If this
turns out to not work so well, I'll switch to using `check` flags
instead.