➕ Generate TS code from GraphQL schema
Use graphql-codegen + several plugins to auto-generate TS types and urql hooks.

I'm choosing to commit the generated file to make it easier to inspect on GitHub and review as part of PRs.

Auto-generate before starting the dev server. Also auto-generate on the host system before starting the build in the Docker container, since there won't be a server running in the container from which to obtain the schema.