🔨 Set compiler options
Set environment-specific compiler options.

We always want `all_warnings: true` so that you get reminded of warnings even when files aren't recompiled.

We ignore module conflicts in dev because it helps when moving files around.

We only compile with warnings as errors in prod; often, we want to test something that temporarily causes a compiler warning, and we don't want that to block us.