✨ Create and show an initial account
Resurrects account view from old app, but trying to do things mock-free now. As such, the various live view tests will be the end-to-end tests.

Given that, there is no longer a need to run tests against the dev or prod instances of the app, so no need for users or multiple accounts.