🔒️ Scope fund fetching to an account
Otherwise, it would have been possible for someone to navigate to a particular fund's show page by its ID without checking that it was their account.

With only a single account in the system, it's not really an issue, but it's better to handle it now than forget about it later.

Also, refactor to extract values from `socket.assigns` at the top of functions to make the remaining code more readable.