✨ Display fund balances
We show the fund balances in the main fund list view, the smaller fund
list sidebar (when viewing a single fund), and in the header for each
single fund page.

We don't actually have balances yet, so for now generate random balances
when necessary.

Balances are stored as a virtual field on each fund.

Use ex_money and ex_money_sql to represent the balance. This library
might be overkill, but seems to be the recommended choice.

Closes #100