Regular withdrawal (#167)
Implements the regular withdrawals feature.

While doing this, I discovered that adding LineItems to an unsaved Transaction doesn't work for more than one line item, so I had to change approach by using `put_assoc` when creating the initial changeset. I refactored the single-fund deposit and withdrawal transactions to match the same pattern.

Reused the `SingleTransactionForm`, which had almost everything needed so renamed it to `TransactionForm`. 

I couldn't find a great way to get the fund names to use as field labels, so using `Enum.at` for now.

I'm using the `action: ignore` feature of Ecto to ignore any line items that have missing or zero amounts. The net result has too much error feedback on the form, but I'm going to live with that for now.

Closes #102