✨ Cover overdrafts when making withdrawals
When making a single- or multi-fund withdrawal, do not allow any fund balances to go negative. Instead, cover the overdraft from the default fund (which is allowed to go negative).

Does not protect against the case where no default fund has been configured. That'll come in a follow-up.

Part of #103.