🚸 Remove account actions from fund show pages
The individual fund page was too cluttered with the extra actions on top, so remove them and replace with a simple account name component.

It was also difficult to cleanly navigate away from a fund's page if the fund became deactivated; this eliminates that problem by making it not possible to do (at least within a single browser tab).

This allowed inlining a bunch of attributes into the newly-renamed AccountBar live component.