Update transaction list on deposit/withdrawal
This was already working due to the fund being updated (current balance), but the transactions weren't showing up in the right spot.

Added `reset: true` to ensure that the new transaction shows up in its proper sort position and used the dom_id from the stream as the id of the table row. I was customizing it, and that was breaking updates.

Also, fixed fund show to only update when its fund changes.