⬆️ Upgrade to Postgres 16.4
Followed the instructions in [this post](https://www.ajfriesen.com/postgresql-upgrades-docker/) to create a new volume and dump/restore the data.

Added a script for installing Postgres with asdf by following [this comment](https://github.com/pythoninthegrass). I remember having trouble with this before, so capturing a script seemed wise.