🐛 Refresh fund transaction list after editing a transaction
When editing the icon, name, or date of a transaction on the fund page, the displayed fund list was not being updated.

We now watch for the `transaction_updated` event and then send an update to the transaction list, causing it to refresh itself.

We also do the same for loan transactions.

Fixes #184