➖ Eliminate all use of Timex
Timex is a wonderful library, but has slowly been getting subsumed into the standard library. Given that, it has not been maintained as thoroughly, so it tends to produce a lot of compiler warnings on newer versions of Elixir. Therefore, we eliminate our dependency on it.

We really only needed a replacement for `Timex.today(:local)`, so we added `LocalTime.today()` instead, which uses NaiveDateTime instead of Timex.

Date.shift is a pretty much direct replacement for Timex.shift.