💄 Adjust column widths in transaction tables
Modify the `.table` component's `:col` slots to take an optional `class` attribute and use it to specify column widths.

Not only does this make the column widths more reasonable, it also keeps them in a fixed position when paginating. Before, the columns would jump around based on the content in a page.